trekk ut hook useIsDesktop
